Tracked commit 1
- Cleanup code
- Optimize startup
- Move ore dictionary registrations into item and block registers and remove ore registers
- Make recipes harder for particle accelerator, quantum assembler, chemical boiler, control rod, turbine, plasma containment block, all hazmat armor recipes, coercion driver, fortron field constructor,  interdiction matrix, biometric identifier and focus matrix recipe
- Make neutron capture chamber operating speed based on reactor temperature. Also make one tritium cell take 90 seconds to produce with a temperature of 4500 C.
- Buff uranium ore count per chunk by 25%
- Made chemical boilers  work as long as they have water for the next tick and not the entire recipe cost.

+++ src/nuclear/physica/nuclear/client/gui/GuiNeutronCaptureChamber.java
@@ -20,7 +20,7 @@ public class GuiNeutronCaptureChamber extends GuiContainerBase<TileNeutronCaptur
 	@Override
 	protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY) {
 		super.drawGuiContainerForegroundLayer(mouseX, mouseY);
-		drawString("Progress: " + (int) ((float) host.getOperatingTicks() * 100 / TileNeutronCaptureChamber.TICKS_REQUIRED) + "%", 86, 73);
+		drawString("Progress: " + (int) (host.getOperatingTicks() * 100 / TileNeutronCaptureChamber.TICKS_REQUIRED) + "%", 86, 73);
 		drawString("Inventory", 8, 73);
 		drawStringCentered(StatCollector.translateToLocal("tile." + NuclearReferences.PREFIX + "neutronCaptureChamber.gui"), xSize / 2, 5);
 	}
